Rename TO=DO LIST to TO-DO LIST
The provided code represents a simple To-Do List application written in HTML and JavaScript. When rendered on a web page, it displays a list of tasks with options to add new tasks, mark tasks as completed, and remove tasks from the list.

Here is a breakdown of the code:

The code begins with HTML markup, including a <head> section with some CSS styles and a <body> section with the page's content.
The <div id="myDIV"> element represents the header of the To-Do List application and contains a title, an input field for adding new tasks, and an "Add" button.
The <ul id="myUL"> element represents the unordered list that displays the tasks. It initially contains some pre-defined list items as examples.
JavaScript code is included within <script> tags. It performs various functions:
Creates a "close" button for each list item and attaches a click event handler to hide the corresponding task when the button is clicked.
Adds a "checked" class to a list item when it is clicked, visually indicating that it has been completed.
Defines the newElement() function, which is called when the "Add" button is clicked. It creates a new list item based on the value entered in the input field and appends it to the list. It also adds a "close" button to the new item and sets up its event handler.